+++ lib/platform/monitor/web.local.sh
@@ -274,9 +274,11 @@ else
 	mv $TEST_RESULT_FILE ~/results/$TM_TARGET_ID/$(hostname -s).xml
 fi
 
-# Clean up the output and lock directories
-rm -rf ~/output/$TM_TARGET_ID
-rm -rf ~/lock/$TM_TARGET_ID
+# Clean up the output and lock directories if on a non-director node
+if [ -z "$(sqlite3 ~/db/cluster.db "SELECT * FROM M7_Nodes WHERE Type='worker' AND Name='$(hostname -s)';")" ]; then
+	rm -rf ~/output/$TM_TARGET_ID
+	rm -rf ~/lock/$TM_TARGET_ID
+fi
 
 # Self destruct the monitor script and destroy the workspace
 rm -rf $M7_TEST_WS
